refactor(roast-timer): 焙煎タイマー関連4ファイルを機能別モジュールに分割
- SetupPanel.tsx (688行) → SetupPanel + ModeSelectView + RecommendedModeView
- useTimerControls.ts (458行) → useTimerControls + useTimerUpdater
- RoastTimer.tsx (402行) → RoastTimer + useRoastTimerDialogs
- useTimerNotifications.ts (372行) → useTimerNotifications + audioUnlocker

Closes #88
