+++ components/roast-timer/SetupPanel.tsx
@@ -4,7 +4,7 @@ import { useState, useEffect, useRef } from 'react';
 import { useAuth } from '@/lib/auth';
 import { useAppData } from '@/hooks/useAppData';
 import { HiPlay } from 'react-icons/hi';
-import { MdTimer, MdLightbulb } from 'react-icons/md';
+import { MdTimer } from 'react-icons/md';
 import { useToastContext } from '@/components/Toast';
 import { ALL_BEANS, type BeanName } from '@/lib/beanConfig';
 import { loadRoastTimerSettings } from '@/lib/roastTimerSettings';
@@ -12,6 +12,8 @@ import { getAllRoastTimerRecords } from '@/lib/roastTimerRecords';
 import { formatTimeAsMinutesAndSeconds, calculateRecommendedTime } from '@/lib/roastTimerUtils';
 import { ROAST_LEVELS, WEIGHTS, DEFAULT_DURATIONS as DEFAULT_DURATION_BY_WEIGHT, type RoastLevel, type Weight } from '@/lib/constants';
 import { convertToHalfWidth, removeNonNumeric } from '@/lib/utils';
+import { ModeSelectView } from './ModeSelectView';
+import { RecommendedModeView } from './RecommendedModeView';
 
 interface SetupPanelProps {
   onStart: (
@@ -310,11 +312,19 @@ export function SetupPanel({ onStart, isLoading }: SetupPanelProps) {
     }
   };
 
-  // 戻るハンドラー
-  const handleBack = () => {
-    setInputMode(null);
-    setRecommendedMode('weight');
-    setBeanName('');
+  // おすすめモード切り替え
+  const handleRecommendedModeChange = (mode: 'weight' | 'history') => {
+    if (mode === 'weight') {
+      setBeanName('');
+      setRoastLevel('');
+      setRecommendedTimeInfo(null);
+    }
+    setRecommendedMode(mode);
+  };
+
+  // 豆名変更
+  const handleBeanNameChange = (value: BeanName | '') => {
+    setBeanName(value);
     setWeight('');
     setRoastLevel('');
   };
@@ -322,83 +332,25 @@ export function SetupPanel({ onStart, isLoading }: SetupPanelProps) {
   // モード選択画面
   if (inputMode === null) {
     return (
-      <div className="flex-1 flex flex-col items-center justify-center px-4 sm:px-6 py-8">
-        {/* タイトルセクション */}
-        <div className="text-center space-y-4 mb-10 sm:mb-12">
-          <div className="inline-flex items-center justify-center w-20 h-20 sm:w-24 sm:h-24 rounded-full bg-gradient-to-br from-amber-400 to-amber-600 shadow-lg mb-3">
-            <MdTimer className="text-white text-4xl sm:text-5xl" />
-          </div>
-          <h3 className="text-3xl sm:text-4xl font-bold text-gray-900 tracking-tight">
-            ローストタイマー
-          </h3>
-          <p className="text-base sm:text-lg text-gray-500">焙煎時間を設定してスタート</p>
-        </div>
-
-        {/* 手動入力フィールド */}
-        <div className="space-y-4 max-w-md mx-auto w-full mb-8 sm:mb-10">
-          <label className="block text-sm sm:text-base font-semibold text-gray-700 uppercase tracking-wide">
-            時間設定
-          </label>
-          <div className="flex gap-3 sm:gap-4">
-            <div className="flex-1">
-              <input
-                type="text"
-                inputMode="numeric"
-                value={durationMinutes}
-                onChange={(e) => handleDurationMinutesChange(e.target.value)}
-                placeholder="分"
-                className="w-full rounded-lg border-2 border-gray-200 px-4 py-3.5 sm:py-4 text-lg sm:text-xl text-gray-900 bg-white focus:border-amber-500 focus:bg-white focus:outline-none focus:ring-2 focus:ring-amber-100 transition-all duration-200 font-semibold text-center min-h-[56px] shadow-sm hover:border-gray-300"
-              />
-            </div>
-            <div className="flex items-end pb-2">
-              <span className="text-3xl font-bold text-gray-400">:</span>
-            </div>
-            <div className="flex-1">
-              <input
-                type="text"
-                inputMode="numeric"
-                value={durationSeconds}
-                onChange={(e) => handleDurationSecondsChange(e.target.value)}
-                placeholder="秒"
-                maxLength={2}
-                className="w-full rounded-lg border-2 border-gray-200 px-4 py-3.5 sm:py-4 text-lg sm:text-xl text-gray-900 bg-white focus:border-amber-500 focus:bg-white focus:outline-none focus:ring-2 focus:ring-amber-100 transition-all duration-200 font-semibold text-center min-h-[56px] shadow-sm hover:border-gray-300"
-              />
-            </div>
-          </div>
-        </div>
-
-        {/* ボタンセクション */}
-        <div className="space-y-3 sm:space-y-4 max-w-md mx-auto w-full">
-          {/* 手動スタートボタン */}
-          <button
-            onClick={handleManualStart}
-            disabled={
-              ((!durationMinutes || durationMinutes.trim() === '') &&
-                (!durationSeconds || durationSeconds.trim() === '')) ||
-              !user ||
-              isLoading
-            }
-            className="w-full flex items-center justify-center gap-2 px-6 py-4 sm:py-5 bg-gradient-to-r from-amber-500 to-amber-600 text-white rounded-xl font-bold text-base sm:text-lg shadow-lg hover:shadow-xl hover:from-amber-600 hover:to-amber-700 active:scale-[0.98] transition-all duration-200 min-h-[60px] sm:min-h-[64px] disabled:from-gray-300 disabled:to-gray-400 disabled:cursor-not-allowed disabled:hover:shadow-lg disabled:active:scale-100 disabled:hover:from-gray-300 disabled:hover:to-gray-400"
-          >
-            <HiPlay className="text-xl sm:text-2xl" />
-            <span>手動でタイマースタート</span>
-          </button>
-
-          {/* おすすめ焙煎ボタン */}
-          <button
-            onClick={() => {
-              setDurationMinutes('');
-              setDurationSeconds('');
-              setRecommendedMode('weight');
-              setInputMode('recommended');
-            }}
-            className="w-full flex items-center justify-center gap-2 px-6 py-4 sm:py-5 bg-gradient-to-r from-amber-50 to-amber-100 text-amber-700 border-2 border-amber-200 rounded-xl font-bold text-base sm:text-lg shadow-sm hover:shadow-md hover:from-amber-100 hover:to-amber-200 hover:border-amber-300 active:scale-[0.98] transition-all duration-200 min-h-[60px] sm:min-h-[64px]"
-          >
-            <MdLightbulb className="text-xl sm:text-2xl text-amber-600" />
-            <span>おすすめタイマー</span>
-          </button>
-        </div>
-      </div>
+      <ModeSelectView
+        durationMinutes={durationMinutes}
+        durationSeconds={durationSeconds}
+        onDurationMinutesChange={handleDurationMinutesChange}
+        onDurationSecondsChange={handleDurationSecondsChange}
+        onManualStart={handleManualStart}
+        onRecommendedMode={() => {
+          setDurationMinutes('');
+          setDurationSeconds('');
+          setRecommendedMode('weight');
+          setInputMode('recommended');
+        }}
+        isStartDisabled={
+          ((!durationMinutes || durationMinutes.trim() === '') &&
+            (!durationSeconds || durationSeconds.trim() === '')) ||
+          !user ||
+          isLoading
+        }
+      />
     );
   }
 
@@ -470,217 +422,22 @@ export function SetupPanel({ onStart, isLoading }: SetupPanelProps) {
 
   // おすすめモード
   return (
-    <div className="flex-1 flex flex-col px-4 sm:px-6">
-      <div className="max-w-md mx-auto w-full flex-1 flex flex-col justify-center py-8">
-        <div className="flex items-center justify-center mb-6 flex-shrink-0">
-          <h3 className="text-xl sm:text-2xl font-bold text-gray-900 flex items-center gap-2">
-            <div className="w-8 h-8 rounded-full bg-gradient-to-br from-amber-400 to-amber-600 flex items-center justify-center">
-              <MdLightbulb className="text-white text-lg" />
-            </div>
-            おすすめタイマー
-          </h3>
-        </div>
-
-        {/* モード切り替えタブ */}
-        <div className="mb-6 flex-shrink-0">
-          <div className="flex gap-2 bg-gray-100 rounded-lg p-1">
-            <button
-              onClick={() => {
-                setRecommendedMode('weight');
-                setBeanName('');
-                setRoastLevel('');
-                setRecommendedTimeInfo(null);
-              }}
-              className={`flex-1 px-4 py-2.5 rounded-md text-sm sm:text-base font-semibold transition-all duration-200 ${
-                recommendedMode === 'weight'
-                  ? 'bg-white text-amber-700 shadow-sm'
-                  : 'text-gray-600 hover:text-gray-900'
-              }`}
-            >
-              重さで設定
-            </button>
-            <button
-              onClick={() => {
-                setRecommendedMode('history');
-              }}
-              className={`flex-1 px-4 py-2.5 rounded-md text-sm sm:text-base font-semibold transition-all duration-200 ${
-                recommendedMode === 'history'
-                  ? 'bg-white text-amber-700 shadow-sm'
-                  : 'text-gray-600 hover:text-gray-900'
-              }`}
-            >
-              過去の記録から設定
-            </button>
-          </div>
-        </div>
-
-        <div className="space-y-5 flex-1">
-          {recommendedMode === 'weight' && (
-            <>
-              <div>
-                <label className="block text-sm font-semibold text-gray-700 uppercase tracking-wide mb-2">
-                  重さ <span className="text-red-500">*</span>
-                </label>
-                <select
-                  value={weight}
-                  onChange={(e) =>
-                    setWeight(e.target.value ? (parseInt(e.target.value, 10) as Weight) : '')
-                  }
-                  className="w-full rounded-xl border-2 border-gray-200 px-4 py-3.5 text-base sm:text-lg text-gray-900 bg-gray-50 focus:border-amber-500 focus:bg-white focus:outline-none focus:ring-4 focus:ring-amber-100 transition-all duration-200 shadow-sm hover:border-gray-300 min-h-[52px]"
-                >
-                  <option value="">選択してください</option>
-                  {WEIGHTS.map((w) => (
-                    <option key={w} value={w}>
-                      {w}g
-                    </option>
-                  ))}
-                </select>
-              </div>
-
-              {weight !== '' && (
-                <div className="bg-gradient-to-br from-amber-50 to-amber-100 border-2 border-amber-200 rounded-xl p-4 shadow-sm">
-                  <p className="text-sm sm:text-base text-gray-700">
-                    重さに応じたおすすめ時間は{' '}
-                    <span className="font-bold text-amber-800">
-                      {weight === 200 ? '8分' : weight === 300 ? '9分' : '10分'}
-                    </span>{' '}
-                    です
-                  </p>
-                </div>
-              )}
-            </>
-          )}
-
-          {recommendedMode === 'history' && (
-            <>
-              <div>
-                <label className="block text-sm font-semibold text-gray-700 uppercase tracking-wide mb-2">
-                  豆の名前 <span className="text-red-500">*</span>
-                </label>
-                <select
-                  value={beanName}
-                  onChange={(e) => {
-                    setBeanName(e.target.value as BeanName);
-                    setWeight('');
-                    setRoastLevel('');
-                  }}
-                  className="w-full rounded-xl border-2 border-gray-200 px-4 py-3.5 text-base sm:text-lg text-gray-900 bg-gray-50 focus:border-amber-500 focus:bg-white focus:outline-none focus:ring-4 focus:ring-amber-100 transition-all duration-200 shadow-sm hover:border-gray-300 min-h-[52px]"
-                >
-                  <option value="">選択してください</option>
-                  {availableBeans.length > 0 ? (
-                    availableBeans.map((bean) => (
-                      <option key={bean} value={bean}>
-                        {bean}
-                      </option>
-                    ))
-                  ) : (
-                    <option value="" disabled>
-                      記録がありません（2件以上の記録が必要です）
-                    </option>
-                  )}
-                </select>
-              </div>
-
-              <div>
-                <label className="block text-sm font-semibold text-gray-700 uppercase tracking-wide mb-2">
-                  焙煎度合い <span className="text-red-500">*</span>
-                </label>
-                <select
-                  value={roastLevel}
-                  onChange={(e) => setRoastLevel(e.target.value as RoastLevel | '')}
-                  disabled={!beanName}
-                  className="w-full rounded-xl border-2 border-gray-200 px-4 py-3.5 text-base sm:text-lg text-gray-900 bg-gray-50 focus:border-amber-500 focus:bg-white focus:outline-none focus:ring-4 focus:ring-amber-100 transition-all duration-200 shadow-sm hover:border-gray-300 min-h-[52px] disabled:bg-gray-100 disabled:cursor-not-allowed disabled:text-gray-400"
-                >
-                  <option value="">選択してください</option>
-                  {availableRoastLevels.length > 0 ? (
-                    availableRoastLevels.map((level) => (
-                      <option key={level} value={level}>
-                        {level}
-                      </option>
-                    ))
-                  ) : beanName ? (
-                    <option value="" disabled>
-                      記録がありません（2件以上の記録が必要です）
-                    </option>
-                  ) : (
-                    <option value="" disabled>
-                      豆を選択してください
-                    </option>
-                  )}
-                </select>
-              </div>
-
-              <div>
-                <label className="block text-sm font-semibold text-gray-700 uppercase tracking-wide mb-2">
-                  重さ <span className="text-red-500">*</span>
-                </label>
-                <select
-                  value={weight}
-                  onChange={(e) =>
-                    setWeight(e.target.value ? (parseInt(e.target.value, 10) as Weight) : '')
-                  }
-                  disabled={!beanName}
-                  className="w-full rounded-xl border-2 border-gray-200 px-4 py-3.5 text-base sm:text-lg text-gray-900 bg-gray-50 focus:border-amber-500 focus:bg-white focus:outline-none focus:ring-4 focus:ring-amber-100 transition-all duration-200 shadow-sm hover:border-gray-300 min-h-[52px] disabled:bg-gray-100 disabled:cursor-not-allowed disabled:text-gray-400"
-                >
-                  <option value="">選択してください</option>
-                  {availableWeights.length > 0 ? (
-                    availableWeights.map((w) => (
-                      <option key={w} value={w}>
-                        {w}g
-                      </option>
-                    ))
-                  ) : beanName ? (
-                    <option value="" disabled>
-                      記録がありません（2件以上の記録が必要です）
-                    </option>
-                  ) : (
-                    <option value="" disabled>
-                      豆を選択してください
-                    </option>
-                  )}
-                </select>
-              </div>
-
-              {!recommendedTimeInfo && beanName && weight !== '' && roastLevel && (
-                <div className="bg-yellow-50 border-2 border-yellow-200 rounded-xl p-4 shadow-sm">
-                  <p className="text-sm sm:text-base text-yellow-800">
-                    この組み合わせの記録が2件未満のため、平均焙煎時間を計算できません。重さに応じたデフォルト時間（
-                    {weight === 200 ? '8分' : weight === 300 ? '9分' : '10分'}）でタイマーを開始します。
-                  </p>
-                </div>
-              )}
-
-              {recommendedTimeInfo && (
-                <div className="bg-gradient-to-br from-amber-50 to-amber-100 border-2 border-amber-200 rounded-xl p-4 shadow-sm">
-                  <p className="text-sm sm:text-base text-gray-700">
-                    過去の記録から、平均焙煎時間は{' '}
-                    <span className="font-bold text-amber-800">
-                      {formatTimeAsMinutesAndSeconds(recommendedTimeInfo.averageDuration)}
-                    </span>
-                    、おすすめタイマー時間は{' '}
-                    <span className="font-bold text-amber-800">
-                      {formatTimeAsMinutesAndSeconds(recommendedTimeInfo.recommendedDuration)}
-                    </span>{' '}
-                    です
-                  </p>
-                </div>
-              )}
-            </>
-          )}
-
-          <div className="pt-2 flex-shrink-0">
-            <button
-              onClick={handleStart}
-              disabled={!weight || (recommendedMode === 'history' && (!beanName || !roastLevel))}
-              className="w-full flex items-center justify-center gap-3 px-6 py-4 bg-gradient-to-r from-amber-500 to-amber-600 text-white rounded-xl font-bold shadow-lg hover:shadow-xl hover:from-amber-600 hover:to-amber-700 active:scale-[0.98] transition-all duration-200 text-base sm:text-lg min-h-[56px] disabled:from-gray-300 disabled:to-gray-400 disabled:cursor-not-allowed disabled:hover:shadow-lg disabled:active:scale-100 disabled:hover:from-gray-300 disabled:hover:to-gray-400"
-            >
-              <HiPlay className="text-2xl" />
-              <span>スタート</span>
-            </button>
-          </div>
-        </div>
-      </div>
-    </div>
+    <RecommendedModeView
+      recommendedMode={recommendedMode}
+      beanName={beanName}
+      weight={weight}
+      roastLevel={roastLevel}
+      availableBeans={availableBeans}
+      availableWeights={availableWeights}
+      availableRoastLevels={availableRoastLevels}
+      recommendedTimeInfo={recommendedTimeInfo}
+      onRecommendedModeChange={handleRecommendedModeChange}
+      onBeanNameChange={handleBeanNameChange}
+      onWeightChange={setWeight}
+      onRoastLevelChange={setRoastLevel}
+      onStart={handleStart}
+      isStartDisabled={!weight || (recommendedMode === 'history' && (!beanName || !roastLevel))}
+    />
   );
 }
